backend/requirements_gemini.txt
# Google Gemini 2.5 Flash Image Generation Requirements
# Install with: pip install -r requirements_gemini.txt

# Google Genai client for Gemini 2.5 Flash Image
google-genai>=0.1.0

# Pillow for image processing
Pillow>=10.0.0

# Additional dependencies for image handling
requests>=2.31.0
